+++ package.json
@@ -25,6 +25,7 @@
     "@supabase/supabase-js": "^2.47.10",
     "@tanstack/react-query": "^5.62.11",
     "@types/react-router-dom": "^5.3.3",
+    "axios": "^1.7.9",
     "jotai": "^2.11.0",
     "nuqs": "^2.3.0",
     "react": "^18.3.1",
