+++ src/contexts/SupabaseProvider.tsx
@@ -1,17 +1,11 @@
-import React, { createContext, ReactNode, useEffect, useState } from 'react';
-import { supabase } from '../apis/supabase';
-import { Session, User } from '@supabase/supabase-js';
-
-type SupabaseContextType = {
-  session: Session | null;
-  user: User | null;
-};
+import { ReactNode, useEffect, useState } from 'react';
+import type { Session, User } from '@supabase/supabase-js';
+import { supabaseAuth } from '../apis/supabase';
+import { SupabaseContext } from '@/contexts/SupabaseContext';
 
 type SupabaseProviderProps = {
   children: ReactNode;
 };
-
-const SupabaseContext = createContext<SupabaseContextType | null>(null);
 
 export function SupabaseProvider({ children }: SupabaseProviderProps) {
   const [session, setSession] = useState<Session | null>(null);
@@ -19,14 +13,14 @@ export function SupabaseProvider({ children }: SupabaseProviderProps) {
 
   useEffect(() => {
     const getSession = async () => {
-      const { data } = await supabase.auth.getSession();
+      const { data } = await supabaseAuth.auth.getSession();
       setSession(data?.session || null);
       setUser(data?.session?.user || null);
     };
 
     getSession();
 
-    const { data: listener } = supabase.auth.onAuthStateChange(
+    const { data: listener } = supabaseAuth.auth.onAuthStateChange(
       (_event, session) => {
         setSession(session);
         setUser(session?.user || null);
@@ -43,13 +37,4 @@ export function SupabaseProvider({ children }: SupabaseProviderProps) {
       {children}
     </SupabaseContext.Provider>
   );
-}
-
-export function useSupabase() {
-  const context = React.useContext(SupabaseContext);
-  if (!context) {
-    throw new Error(
-      '❌useSupabase를 사용하려면 SupabaseProvider의 내부에 작성해야 합니다.',
-    );
-  }
 }
